+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Method `isFirstDayOfMonth()` - Checks if `DateTime` is in the first day of a month.
+* Method `isLastDayOfMonth()` - Checks if `DateTime` is in the last day of a month.
 
 ## [0.2.0] - 2020-11-18